SellmyShow_Phase1-main/SellmyShow_V1_0/ticket_seller_backend/requirements.txt
Flask==3.1.0
flask-cors==6.0.1
pymongo==4.10.1
PyJWT==2.10.1
razorpay==1.4.2
gunicorn>=20.1.0
python-dotenv==1.1.1
pillow==11.2.1
pytesseract==0.3.13

google-generativeai==0.8.5
google-api-core==2.25.1
google-api-python-client==2.176.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
googleapis-common-protos==1.70.0
googlemaps==4.10.0

requests==2.32.3

bcrypt==4.3.0 